Remove comment about inefficiency
Algorithm `filter_by_time` could be done in O(N log N) by sorting
records and after that use binary search, whereas now the algorithm is
O(N) by doing a direct linear search. As the number of records is small,
this very likely is not going to be a problem. To be implemented later
on, if performance problem arises.